backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
celery>=5.3,<6.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
django-storages>=1.14,<2.0
boto3>=1.34,<2.0
python-dotenv>=1.0,<2.0
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0
